+++ src/Entity/GenericManager.php
@@ -19,6 +19,7 @@ namespace Gpupo\CommonSdk\Entity;
 
 use Gpupo\Common\Entity\CollectionInterface;
 use Gpupo\CommonSdk\Map;
+use Closure;
 
 class GenericManager extends ManagerAbstract
 {
@@ -39,7 +40,7 @@ class GenericManager extends ManagerAbstract
         return $perform->getData();
     }
 
-    public function requestWithCache(array $route, string $identifier, $body = null, bool $renew = false): CollectionInterface
+    public function requestWithCache(array $route, string $identifier, $body = null, bool $renew = false, Closure $normalizer = null): CollectionInterface
     {
         $cacheId = $this->getClient()->simpleCacheGenerateId($identifier);
 
@@ -53,12 +54,18 @@ class GenericManager extends ManagerAbstract
         }
 
         $response = $this->getFromRoute($route, $this->getClient()->getOptions()->toArray(), $body);
+
+        if ($normalizer) {
+            $response = $normalizer($response);
+        }
+
         $this->getClient()->getSimpleCache()->set($cacheId, $response, $this->getClient()->getOptions()->get('cacheTTL', 3600));
         $this->log('info', 'Saving response to cache', [
             'route' => $route,
             'cacheId' => $cacheId,
             'renew' => $renew,
         ]);
+
 
         return $response;
     }